+++ NEWS.md
@@ -48,3 +48,9 @@
 # NADA2 1.1.3
 
 * Minor bug fix for `cencorreg.R` function (2023-01-31)
+
+# NADA2 1.2.0
+
+* Minor change to `censeaken.R` and `kenplot.R` to allow for some customization of plot using ... in the function (2023-04-13)
+
+* Minor change to `cboxplot.R` to fix factor grouping, added detection limit lty and col arguments (2023-04-13)